Correction

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -7,7 +7,7 @@
 
 size_t print_list(const list_t *h)
 {
-	size_t number_of_elements;
+	int number_of_elements;
 
 	number_of_elements = 0;
